pyproject.toml
[project]
name = "otel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-monitor-opentelemetry>=1.8.7",
    "flask>=3.1.3",
    "opentelemetry-distro>=0.61b0",
    "pip>=26.0.1",
]
